feat: create GET methods for / and /health routes
This commit introduces a basic Express implementation
with GET methods for / (returns a JSON "Hello" message)
and for /health (returns a JSON timestamp and OK status).

It also introduces a basic Dockerfile usage for the sake
of reproduceability. Further commits will include
instructions to run through Docker as well as through a
local setup.